refactor: rename connector package to span_metrics and expose __version__
Rename the cloudwatch/connector package to cloudwatch/span_metrics
(inner span_metrics_connector.py -> connector.py), update the entry
point and all imports, expose __version__ from the top-level package
(OTel convention), and drop the redundant SpanMetricsConnector._lib_version
cached copy in favor of __version__ directly.
