+++ cloudwatch-plugin-otel/src/plugins/opentelemetry/cloudwatch/__init__.py
@@ -1,7 +1,8 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
-from plugins.opentelemetry.cloudwatch.connector.instrumentor import SpanMetricsInstrumentor
-from plugins.opentelemetry.cloudwatch.connector.span_metrics_connector import SpanMetricsConnector
 from plugins.opentelemetry.cloudwatch.sampler.always_record_sampler import AlwaysRecordSampler
+from plugins.opentelemetry.cloudwatch.span_metrics.connector import SpanMetricsConnector
+from plugins.opentelemetry.cloudwatch.span_metrics.instrumentor import SpanMetricsInstrumentor
+from plugins.opentelemetry.cloudwatch.version import __version__
 
-__all__ = ["AlwaysRecordSampler", "SpanMetricsConnector", "SpanMetricsInstrumentor"]
+__all__ = ["AlwaysRecordSampler", "SpanMetricsConnector", "SpanMetricsInstrumentor", "__version__"]
