+++ cloudwatch-plugin-otel/src/plugins/opentelemetry/cloudwatch/span_metrics/__init__.py
@@ -1,5 +1,5 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
-from plugins.opentelemetry.cloudwatch.connector.span_metrics_connector import SpanMetricsConnector
+from plugins.opentelemetry.cloudwatch.span_metrics.connector import SpanMetricsConnector
 
 __all__ = ["SpanMetricsConnector"]
